requirements.txt
Django>=5.0,<5.1
psycopg[binary]>=3.1 
python-dotenv>=1.0 
channels>=4.0
daphne>=4.1 
redis>=5.0
celery>=5.3
Pillow>=10.2  
qrcode[pil]>=7.4  
djangorestframework==3.15.1
django-cors-headers==4.3.1